requirements.txt
openai
langchain
langchain-community
faiss-cpu
python-dotenv
langchain-openai
streamlit
pandas
tqdm
tiktoken
pinecone
langchain-pinecone
